:construction: Deleted consumer and tradesman model in favor of one universal user model
