+++ libs/redux_comp/lib/actions/login_action.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 // import 'package:amplify/models/Advert.dart';
 // import 'package:amplify/models/Consumer.dart';
 // // import 'package:amplify_api/amplify_api.dart';
-// import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
+import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
 import 'package:redux_comp/app_state.dart';
@@ -18,12 +18,15 @@ class LoginAction extends ReduxAction<AppState> {
   @override
   Future<AppState?> reduce() async {
     try {
-      /*SignInResult res = */await Amplify.Auth.signIn(
+      SignInResult res = await Amplify.Auth.signIn(
         username: email,
         password: password,
       );
-      /*String id =*/ (await Amplify.Auth.getCurrentUser()).userId;
-      // return state.replace(user: ConsumerModel());
+
+      AuthUser user = await Amplify.Auth.getCurrentUser();
+      String id = user.userId;
+      List<AuthUserAttribute> userAttr = await Amplify.Auth.fetchUserAttributes();
+      // await Amplify.Auth.signOut();
       return state;
       // exception will be handled later
       // } on AuthException catch (e) {
