+++ libs/redux_comp/lib/actions/register_user_action.dart
@@ -1,13 +1,7 @@
-// import 'package:amplify/amplify.dart';
-// import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-// import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
 import 'package:flutter/foundation.dart';
-import 'package:redux_comp/models/user_models/consumer_model.dart';
-// import 'package:flutter/foundation.dart';
-// import 'package:redux_comp/models/user_models/consumer_model.dart';
 import '../app_state.dart';
 import '../models/user_models/partial_user_model.dart';
 
@@ -17,15 +11,16 @@ class RegisterUserAction extends ReduxAction<AppState> {
   final String cellNo;
   final String location;
   final String password;
+  final String userType;
 
-  RegisterUserAction(this.username,this.name,this.cellNo,this.location, this.password);
+  RegisterUserAction(this.username,this.name,this.cellNo,this.location, this.password, this.userType);
 
   @override
   Future<AppState?> reduce() async {
     try {
       Map<CognitoUserAttributeKey, String> userAttributes = {
         CognitoUserAttributeKey.email : username,
-        CognitoUserAttributeKey.phoneNumber: cellNo
+        CognitoUserAttributeKey.familyName: userType
       };
 
        SignUpResult res = await Amplify.Auth.signUp(
