+++ libs/redux_comp/lib/actions/verify_user_action.dart
@@ -1,10 +1,11 @@
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:redux_comp/models/user_models/consumer_model.dart';
 import 'package:redux_comp/models/user_models/partial_user_model.dart';
 
 import '../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
+
+import '../models/user_models/user_model.dart';
 
 // import '../models/user_models/consumer_model.dart';
 
@@ -21,7 +22,7 @@ class VerifyUserAction extends ReduxAction<AppState> {
 
       if (res.nextStep.signUpStep == "DONE") {
         return state.replace(
-          user: ConsumerModel(state.user!.getId(), state.user!.getName(), state.user!.getEmail(), res.nextStep.signUpStep)
+          user: UserModel(state.user!.getId(), state.user!.getEmail(), res.nextStep.signUpStep)
         );
       } else print(res.nextStep.signUpStep);
        return state;// .replace(
