--- libs/redux_comp/lib/models/user_model.dart
@@ -1,40 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:redux_comp/models/user_models/consumer_model.dart';
-import 'package:redux_comp/models/user_models/tradesman_model.dart';
-
-@immutable
-abstract class UserModel {
-  final String _id;
-  final String _name;
-  final String _email;
-  final String _confirmed;
-
-  const UserModel(this._id, this._name, this._email, this._confirmed);
-
-  String getId() {
-    return _id;
-  }
-
-  String getName() {
-    return _name;
-  }
-
-  String getEmail() {
-    return _email;
-  }
-
-  String getConfrimed() {
-    return _confirmed;
-  }
-
-  UserModel replace();
-
-  // convert into required class
-  ConsumerModel asConsumer() {
-    return this as ConsumerModel;
-  }
-
-  TradesmanModel asTradesman() {
-    return this as TradesmanModel;
-  }
-}
